requirements.txt
numpy==1.19.5
streamlit==0.82.0
tensorflow-cpu==2.5.0
keras==2.4.3
pillow==8.2.0
